requirements.txt
youtube-transcript-api
langchain-community
faiss-cpu
tiktoken
python-dotenv
langchain
langchain-huggingface 
sentence-transformers
tf-keras
